Introduce simple ASM generation
Currently, only `OP_NOP`, `OP_PUSH8`, `OP_ADD`, and `OP_EXIT` are
supported (the minimal set of instructions to support exiting with
some exit code and accessing the stack). Note: `NOP`is "supported" in
the sense that it is completely ignored by the generator, rather than
issuing a "not implemented" error message.