Add lexing for integer literal prefixes and suffixes
The prefixes are `0x-` for hexadecimal and `0b-` for binary (or no
prefix for decimal). The suffixes are `-w` for words, `-t` for bytes,
`-s8`, `-s16`, `-s32` and `-u8`, `-u16`, `-u32` for signed/unsigned
fixed width integers. A suffix-less literal is of type `int` (signed,
word-sized)

Additionally, the literal may be preceded by a sign (`-` or `+`).
Unlike in many other languages, the sign is part of the literal.