Consume integer literal prefix while parsing it
A new struct is returned, containing both the sign (`\0`, `-` or `+`)
and the base (10, 16 or 2). The `parse_integer_prefix()` function is
passed a string view containing the integer, which it modifies to
consume any prefix. The function assumes the literal is
well-formed (isn't invalid like `0x` or `h67`), which it should be if
it was returned by the lexer.