Lex character literals like strings to get better error messages
Previously, `'hi'` would be an "unterminated character literal",
because the lexer would only see `'h` followed by something that's not
`'`. Now, the lexer consumes characters until it finds `'`. It also
keeps track of how many characters have been read and rejects literals
without a length of 1.

In the future, the permissible range will
increase to 1-4 to allow for UTF-8 literals (but the encoding won't be
validated until parsing, so `'bad'` will be accepted by the lexer but
rejected by the parser. Right now, though, only single-byte
character literals are allowed.