(Re)Allow byte type for `print-char`
Before the `char` type was introduced, `print-char` expected a byte.
Since we currently have no way of (a) reading a UTF-8 character from a
string, and (b) converting a value of type `byte` to type `char`, we
have to allow `print-char` to take `byte`s if we want to support
old-style string printing. This could be re-visited in the
future. Perhaps a print-a-single-byte-as-part-of-UTF8 instruction.