Expand stubs for CALLn and RET instructions
In time, these will cause the current `block` pointed to by the
interpreter to change to taht of the current function. However, this
requires filoe level code to be saved as an actual function. In time,
this will actually be the `main()` function, but for now we'll just
make "function 0" special.