Only emit RET when it's needed
If the user already used the `ret` keyword, we don't want to emit two
RETs. Of course, we also need to make sure that the would-be final
instruction is not a jump destination (otherwise we'd jump out of the
function and into the abyss (in the interpreter; in the generated
assembly we could fall through to the next function (if it exists), or
just fall off the end of the executable and crash).