Introduce `PRI_SV` and `SV_FMT` for string views in printf et al
The way these macros work work is sneaky (and perhaps a little hacky,
but it's within the realms of C's macro system). PRI_SV is just a
wrapper around `(%).*s` (the `%` is not included). SV_FMT unpacks the
string view in the correct order for the format specifier. It
truncates overlong strings to `INT_MAX` length. This should be plenty
long enough, anyway, but just to be safe we add the check.