Use new string view printf macros for printing symbol names
I know I changed these only 2 commits ago, but I changed my mind
again. After all, who needs a string that has a length greater than
INT_MAX anyway? Is that even possible? Even if so, it's probably not
the sort of thing we want to be copying around (and probably wouldn't
fit in the region, anyway).