Add TODO comments to remember the signedness issue
Right now, the operations assume that the arguments are signed. This
doesn't affect equality and non-equality, but the other inequlaities
can give unexpected results for large unsigned values. The best way to
fix this is to introduce separate instructions for signed/unsigned
versions of the inequalities. Of course, this is only in the WIR.