Add Python module for working with Bude's WIR
This exists to make it easy to interface with the WIR code in Python.
There are come definitions encapsulating the IR code. In particular,
there is a Python enum which corrsponds to the C enum `w_opcode` in
`ir.h`. It is important to keep these two enums in agreement. There is
also a class encapsulting the IR code itself, supporting iteration
over the block, which correctly chunks the bytes into opcodes and any
operands. Instructions themselves (opcode + operands) are also
encapsulated, mainly to support custom printing (S-expression--like).