Add `decomp` before `pop` in old-style string handling examples
The issue is that when the examples were written, strings pushed two
values to the stack, so many examples simply popped the length
straight away since it wasn't needed. Now that strings count as a
comp, popping after pushing a string pops the whole comp. Using
`decomp` after a string literal decomposes the comp back to the
pointer and length that the previous code expected, which is easier
than updating the code in the "idiomatic" way with comp fields.