Introduce floating-point literals in the lexer.
The compiler doesn't do anything with them yet, but they lex
correctly.

This also necessitated a change to how integer literals are lexed. A
brief rundown is:
  token starting with digit, '+', '-', '.':
    consume any '+' or '-' then check first character:
      not '0':
        lex as a decimal number (int or float)
      '0'; check next character:
        'x':
          lex as hexadecimal (int only)
        'b':
          lex as binary (int only)
         _:
          lex as decimal (int or float)

When lexing a decimal number, we start by assuming it's an integer and
lex appropriately. If we see a '.', we assume it's a float instead and
lex the fractional part. If we then see an 'e' (regardless of whether
we saw a '.'), we again assume float and lex the exponent (including
optional sign). Depending on the assumptions we made, we try to lex
either an integr suffix or a float suffix.

There is a slight added complexity since we allow float literals of
the form 42f32, which has no '.' or 'e' to signify it's being a
float. Instead, when we lex the suffix of an assumed integer, we peek
at the next character. If it's 'f', we lex it as a float suffix instead.