Add `PRINT_FLOAT` instruction for priting double-precision floats
The instruction `FPROM` is also required to promote f32 to f64.
This is complemented by `FPROML` (lhs promotion) and `FDEM` (demote
f64 to f32).

This promotion is applied by the type checker before emitting the
`PRINT_FLOAT` instruction, allowing that instrucion to work on doubles
only (which matches `printf` in C, since float varargs are promoted to
double). This technique of promoting before the print is copied from
`PRINT_INT`, which works the same way.