Fix out-of-bounds memory access bug in `grow_block()`
We didn't reallocate the `locations` array, so new locations were
written past the end of the original allocaion. This has now been
fixed.

It only caused issues when an example program got big enough, which
only seemed to happen with the float.bude example (probably due to the
larger PUSH instructions from all the f64s).