Add rest of float arithmetic operations
Unlike for integers, with floats, we need to have a separate operation
for each size/precision (luckily there are only two). If we didn't do
this, what would even be the point of f32 if it always needed to be
converted to f64 for operations? Instead, each operation gets a pair
of instructions suffixed with -F32 and -F64 as appropriate.

As part of this, we have introduced the 'float conversions' table,
which is similar to the 'arithemtic conversions' table, but for when
the operands are floats instead. It handles promoting the floats to
the right type. If we mix f32 and f64 in an arithmetiuc operation, the
f32 gets promoted to f64, which is also the result type. If both are
f32s, we stay in the f32 world.

Currently, mixing integers and floating-points is not supported, but
it will be in the future, with the conversion rules being that the
integer is promoted to word size (64 bits) then converted to the
floating-point type. The result is of the same floating-point type.

To summarise: f32 f32 -> f32; f32 in -> f32; f64 _ -> f64.