Introduce 'to'-conversions with type checking
'To'-conversions are actual value-preserving conversions between
types. They can be contrasted with 'as'-conversions which simply
reinterpret the bit representations.

The general rule for conversions involving integers is to first
promote to a word-size integer, then truncate as necessary. For
integer--floating-point conversions, the conversion happens through
the int type. That is, one can convert any type to `int` and then
`int` to any type. For character types, a second 'middle-ground' type
is used, representing a UTF-32 codepoint. This is very similar to a
`u32` (although strictly has a narrower range) and can converted
from/to the `char` type, which is encoded as UTF-8.