Ensure there's enough space in byte array to interpret `PRINT_CHAR`
This code relies on the zeros after the character to terminate the
string. Previously, the bytes were stored in a 4-array, which would
not be null-terminated in the case of a 4-byte UTF-8 encoding, which
would cause a buffer overrun. Instead, we just use an 8-byte array
since we know at least the last 4 bytes will be zeroed (due to Bude's
empty space rules).