Change `decode_utf8()` to take a start and end pointer
The start pointer is a pointer to const char, and points to the start
of the UTF-8 bytes to be read. The end pointer is a pointer to pointer
to const char. The pointer it points to will be set to point to the
next byte to be read in the UTF-8 stream (the first byte _after_ the
current codepoint). This parameter can be set to NULL, in which case
it is ignored.

Previously, this was all done through one pointer, which acted like
the `end` pointer. The reason for this change was to remove the need
for a dummy variable to hold the pointer if you only need to read one
character (for instance from a `char` value on the stack). There is a
slight change in behaviour, however. Now, the end pointer will only be
written to if the UTF-8 bytes were decoded successfully.