Use X macros for opcode lists
We define macros for T_OPCODES and W_OPCODES, which consist of the
identifiers of each opcode as an argument to an X macro. Then, we can
use one X maro to define the enum (just the opcode name followed by a
comma) and another X macro to define the array of opcodes (using the
opcode name as a designator and the stringised opcode name as a value).