Ensure loop counter is restored after skipping a loop in codegen
The way the code was written previously, the loop counter would be
pushed to the stack before the loop condition was checked. In the case
of skipped loops, this meant that the current loop counter was lost.
Instead, now we only push the old loop counter when we know we need to
enter the loop; if we skip the loop, rdi (which holds the loop
counter) remains unchanged.