Use stack frames on aux in codegen to manage local loop stacks
Each function call gets its own loop stack instead of the previous
global one. To achieve this, we effectively have stack frames for the
auxiliary stack. When a function is called, the return address is
moved to aux and then the current contents of the rbx register is
pushed to it. This register is then updated with the current value of
rsi (the auxiliary stack pointer), acting as a frame pointer. Locals
can then live on the space above rbx. Every loop variable lives at
some offset from rbx which is known at compile time. When a function
with a loop is called from within a loop in the caller, the callee
saves the caller's rdi loop counter in its own loop stack. This
process goes all the way down, ensuring that rdi is always preserved
across function calls without needing any special handling at the callsite.