Add test case for a loop containing a recursive call
Making this work in the codegen required a rewrite of how for loops
were handled. Before, we assumed there was an upper bound on the depth
of for loops in a program known at compile time. This was true when
there was only one function, but it breaks down when we can call a
function from within a loop (doesn't even have to be recursive).

To fix this, we essentially use a stack frame on aux which holds the
'loop stack' of that function call. Within a function call, we know
the maximum for loop depth, which is the same across all calls to the
function and known at compile time, so we reserve that amount of space
on aux when we enter the function.