Introduce helper macros for working with dynamic arrays
These assume the argument "`da`" is a pointer to a structure
with the members `count`, `capacity` and `items`. The `items` member
is a pointer to the array items and should be given the relevant
type (the macro is type-agnostic but uses `sizeof` so the type must be
a known in the dynamic array struct itself.