Bump BWF version to 2
This is a backwards-imcompatible change to the format introducing a
new field `data-info-field-count` into the `DATA-INFO` section. This
field allows for future changes to be backwards-compatible with
version 2, since we can use the count to skip over unknown fields. To
keep things principled, new fields shall only be added below existing
ones within each section.