Introduce BudeBWF version 3 to have an entry-size field
This is a minor update to version 2 which adds an extra field before
the size field for a function table entry. This means that reading a
version 3 file in version 2 will lead to 4 bytes of garbage at the
start of each function, which might cause issues but doesn't result in
data losses as such.