Introduce `WRITE()` and `WRITE_OR_ERR()` macros
The former expands to a call to `fwrite()` and takes an object as the
first parameter which must support `sizeof`. The latter is sugar for
the common idiom of `if (fwrite(...) != 1) return err_ret;`, taking
error-case return value as the thrid parameter. While this hides the
control flow somewhat, it brings the action of writing to the
forefront, and the hidden control flow is only an early exit in the
case of an error, anyway.