Compute jump info table after loading a function from BudeBWF
This means we don't need to add the jump info table to the BWF
format (at least for now). There are some good arguments to add it to
the format, but, on the other hand, the information is readily
computed by just walking the bytecode and adding any jump destinations
to the table, so we'll keep it simple for now.