Support external calls in the type checker
This is essentially the same as type checking normal function
calls. The only difference is what type of object holds the
`siganture` object, so the code to type check the signature has been
extracted to a separate function, with the old `check_function_call()`
function now just acting as a wrapper around this.