Add functions to save/restore certain points in a region
You can call `record_region()` to get an opaque `REGION_RESTORE` type
which can be passed to `restore_region()` to restore the region to the
point at which `record_region()` was called. It is undefined behaviour
to call `restore_region()` with a later retsore point after calling it
with an earlier one.