Don't prompt the user to "use `bude --help`" if they used `--help`
We customise the message at the end of the explanation based on
whether the user specified the `--help` option. If they did, we refer
them to the help message (which should be printed above) and note that
the option that caused the message to be printed was `--help` (aka
`-h`, `-?`). Otherwise, prompt them to use the `--help` option for
more information.